backend_ws/requirements.txt
fastapi
uvicorn
websockets
paho-mqtt
python-multipart
